requirements.txt
streamlit
transformers
torch
faiss-cpu
sentence-transformers
python-docx
PyMuPDF
packaging
